refactor(ui): align with VS Code structure — ActivityBar / Sidebar / approval-in-sidebar
Naming refactor + structural cleanup. Drops the failed PanelGroup
experiment in favor of the actual VS Code layout model.

Renames:
  Sidebar.tsx          → ActivityBar.tsx (the leftmost icon column)
  SecondarySidebar.tsx → Sidebar.tsx     (the activity-specific navigator)

New term mapping (VS Code official):
  ActivityBar = leftmost vertical icon column
  Sidebar     = the panel next to ActivityBar showing activity-specific content
  EditorPart  = the main editor area (not yet introduced — single-editor for now)
  Panel       = the BOTTOM dock (Terminal/Output/Problems) — not used yet

PushApprovalPanel ("Trading as Git") moved out of ChatPage. It's an
ACTION panel (like Remix's Deploy & Run), not editor content — so it
belongs in the left Sidebar alongside the channel list. New
ChatSidebarContent stacks: channel list (top) + approval panel (bottom).

ChatPage is now plain content rendered inside the editor area —
no PanelGroup wrap, no react-resizable-panels dep. The initial-load
flash that fought us across multiple attempts disappears as a side
effect (PanelGroup was the cause; removing it removes the flash).

Sidebar.tsx body switched from `flex-1 overflow-y-auto` to
`flex-1 min-h-0 flex flex-col` so multi-section content (channels +
approval) can manage its own scroll regions. PushApprovalPanel sheds
its right-panel-only `w-72 shrink-0 border-l` styling — now fills
whatever container it lands in.

react-resizable-panels removed from deps. Will revisit once we have
a real reason for split editor groups (and time to integrate carefully).
