+++ ui/package.json
@@ -16,7 +16,6 @@
     "marked-highlight": "^2.2.1",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "react-resizable-panels": "^2.1.9",
     "react-router-dom": "^7.13.1",
     "recharts": "^3.8.0"
   },
