+++ ui/src/App.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 import { Routes, Route, Navigate, useLocation } from 'react-router-dom'
+import { ActivityBar } from './components/ActivityBar'
 import { Sidebar } from './components/Sidebar'
-import { SecondarySidebar } from './components/SecondarySidebar'
 import { ChannelConfigModal } from './components/ChannelConfigModal'
 import { ChannelsProvider, useChannels } from './contexts/ChannelsContext'
 import { SECTIONS, STANDALONE_ROUTES, REDIRECT_ROUTES, findActiveSection } from './sections'
@@ -43,15 +43,15 @@ function AppShell() {
 
   return (
     <div className="flex h-full">
-      <Sidebar open={sidebarOpen} onClose={() => setSidebarOpen(false)} />
+      <ActivityBar open={sidebarOpen} onClose={() => setSidebarOpen(false)} />
 
       {section && (
-        <SecondarySidebar
+        <Sidebar
           title={section.title}
           actions={section.Actions ? <section.Actions /> : undefined}
         >
           <section.Secondary />
-        </SecondarySidebar>
+        </Sidebar>
       )}
 
       <main className="flex-1 flex flex-col min-w-0 min-h-0 bg-bg">
